confirmを実装する

+++ frontend/src/hooks/useDeactivateAccountButtonAction.tsx
@@ -9,10 +9,14 @@ export const useDeactivateAccountButtonAction = () => {
   const [currentUser, setCurrentUser] = useRecoilState(currentUserState);
   const navigate = useNavigate();
   const onCLickDeactivateAccountAction: React.MouseEventHandler<HTMLButtonElement> = (e) => {
-    client.delete(`users/${currentUser.user.id}`).then(response => {
-      response.data.status === 'delete' && navigate('/', {state: {message: '退会しました', type: 'success-message', condition: true}} );
-      response.data.status === 'delete' && setLoginStatus({status: false}) && setCurrentUser({user:{}});
-    });
+    if(window.confirm('本当に退会しますか？')){
+      client.delete(`users/${currentUser.user.id}`).then(response => {
+        response.data.status === 'delete' && navigate('/', {state: {message: '退会しました', type: 'success-message', condition: true}} );
+        response.data.status === 'delete' && setLoginStatus({status: false}) && setCurrentUser({user:{}});
+      });
+    } else {
+      window.alert('退会処理がキャンセルされました');
+    }
   };
 
   return { onCLickDeactivateAccountAction, loginStatus };
